things/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-autoslug==1.9.8
django-cities-light==3.9.2
django-model-utils==4.3.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
phonenumbers==8.13.0
Pillow==9.3.0
progressbar2==4.2.0
psycopg2==2.9.5
python-utils==3.4.5
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tqdm==4.64.1
Unidecode==1.3.6
